+++ frontend-backoffice/package.json
@@ -15,7 +15,7 @@
     "vue": "^2.6.11",
     "vue-axios": "^3.2.4",
     "vue-data-tables": "^3.4.5",
-    "vue-router": "^3.2.0"
+    "vue-router": "^3.2.0",
     "vuex": "^3.6.2"
   },
   "devDependencies": {
