+++ frontend-backoffice/src/App.vue
@@ -7,12 +7,9 @@
   </div>
 </template>
 <script>
-
 import Navbar from "./components/Navbar.vue";
-import { store } from "./store/store";
 
 export default {
-  store,
   name: "app",
   components: { Navbar },
 };
